Refactor Codex cache facade
