+++ src/cli/install-codex/codex-cache-mcp-manifest.ts
@@ -1,0 +1,61 @@
+import { readFile, writeFile } from "node:fs/promises"
+import { join, sep } from "node:path"
+import { resolveBundledMcpRuntimeArg } from "./codex-cache-bundled-mcps"
+import { exists, isRecord } from "./codex-cache-fs"
+import { resolveCachedRuntimePath } from "./codex-cache-paths"
+
+export async function rewriteCachedMcpManifest(pluginRoot: string, sourceRoot = pluginRoot): Promise<void> {
+  const manifestPath = join(pluginRoot, ".mcp.json")
+  if (!(await exists(manifestPath))) return
+  const raw = await readFile(manifestPath, "utf8")
+  const parsed: unknown = JSON.parse(raw)
+  if (!isRecord(parsed) || !isRecord(parsed.mcpServers)) return
+  let changed = false
+  for (const server of Object.values(parsed.mcpServers)) {
+    if (!isRecord(server)) continue
+    if (server.cwd === "." || server.cwd === "./") {
+      delete server.cwd
+      changed = true
+    }
+    const currentArgs = server.args
+    if (!Array.isArray(currentArgs)) continue
+    const nextArgs = currentArgs.map((arg) => {
+      if (typeof arg !== "string") return arg
+      const bundledMcpRuntimeArg = resolveBundledMcpRuntimeArg(pluginRoot, arg)
+      if (bundledMcpRuntimeArg !== null) return bundledMcpRuntimeArg
+      if (arg.startsWith("./") || arg.startsWith("../")) return resolveCachedRuntimePath(pluginRoot, sourceRoot, arg)
+      return arg
+    })
+    if (nextArgs.some((value, index) => value !== currentArgs[index])) {
+      server.args = nextArgs
+      changed = true
+    }
+  }
+  if (changed) await writeFile(manifestPath, `${JSON.stringify(parsed, null, "\t")}\n`)
+}
+
+export async function rewriteCachedManifestRoot(pluginRoot: string, fromRoot: string, toRoot: string): Promise<void> {
+  const manifestPath = join(pluginRoot, ".mcp.json")
+  if (!(await exists(manifestPath))) return
+  const raw = await readFile(manifestPath, "utf8")
+  const parsed: unknown = JSON.parse(raw)
+  if (!isRecord(parsed) || !isRecord(parsed.mcpServers)) return
+  let changed = false
+  for (const server of Object.values(parsed.mcpServers)) {
+    if (!isRecord(server)) continue
+    const currentArgs = server.args
+    if (!Array.isArray(currentArgs)) continue
+    const nextArgs = currentArgs.map((arg) => {
+      if (typeof arg !== "string") return arg
+      if (arg === fromRoot) return toRoot
+      const prefix = `${fromRoot}${sep}`
+      if (!arg.startsWith(prefix)) return arg
+      return `${toRoot}${arg.slice(fromRoot.length)}`
+    })
+    if (nextArgs.some((value, index) => value !== currentArgs[index])) {
+      server.args = nextArgs
+      changed = true
+    }
+  }
+  if (changed) await writeFile(manifestPath, `${JSON.stringify(parsed, null, "\t")}\n`)
+}
